expand BitInputStream to allow reading of up to 63 bits at once
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1654638 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/compress/compressors/z/ZCompressorInputStream.java
@@ -40,9 +40,9 @@ public class ZCompressorInputStream extends LZWInputStream {
     
     public ZCompressorInputStream(InputStream inputStream) throws IOException {
         super(inputStream, ByteOrder.LITTLE_ENDIAN);
-        int firstByte = in.readBits(8);
-        int secondByte = in.readBits(8);
-        int thirdByte = in.readBits(8);
+        int firstByte = (int) in.readBits(8);
+        int secondByte = (int) in.readBits(8);
+        int thirdByte = (int) in.readBits(8);
         if (firstByte != MAGIC_1 || secondByte != MAGIC_2 || thirdByte < 0) {
             throw new IOException("Input is not in .Z format");
         }
